Implement q_insert_head
Insert an element with the value char *s at the head of the queue.
Allocate memory for the new element and its value, copy s to the element,
and use list_add to insert the element at the head. If the head or s is
NULL, return false. If any allocation fails, return false.

Change-Id: Idee45783be1ed936010fecc222de844164b5d4c1