Implement q_insert_tail
Insert an element with the value char *s at the tail of the queue.
Allocate memory for the new element and its value, copy s to the element,
and use list_add_tail to insert the element at the tail. If the head or s is
NULL, return false. If any allocation fails, return false.

Change-Id: Ic9c3d9bc552d3e2750775f559a34d3f509757c2b