Implement q_remove_head
Remove an element from the head of the queue and copy its value to char *sp.
Fetch the address of the element from the head of the queue. If sp is not
NULL, use strncpy to copy the element's value to sp. Finally, remove
the node using list_del and return the address of the element. If the head
is NULL or the queue is empty, return NULL.

Change-Id: I57936f1e775d28b135f6484e7e167c315dc17c5e