Implement q_remove_tail
Remove an element from the tail of the queue and copy its value to char *sp.
Fetch the address of the element from the tail of the queue. If sp is not
NULL, use strncpy to copy the element's value to sp. Finally, remove
the node using list_del and return the address of the element. If the head
is NULL or the queue is empty, return NULL.

Change-Id: I03ccc2a6df99d25a44aa5554cbfae6c125a0a847