Implement q_delete_mid
Delete the middle node in the queue. If there is only a single node in the
queue, delete the node. When there are multiple nodes, use fast and slow
pointers to find the middle node and delete it. If the node is deleted
successfully, return true. If the head is NULL or the queue is empty,
return false.

Change-Id: I0dcad6289f8a6250996b623b2337dbe71f5d1917