Implement q_delete_dup
Delete all neighboring nodes that have duplicate values. Check whether a
node's value is the same as the next or previous node. If it is the same,
delete the node; otherwise, iterate to the next node. If all duplicate nodes
are deleted, return true. If the head is NULL or the queue is empty,
return false. If there is only one node in the queue, do nothing and
return true.

Change-Id: I74efbc94abc8ed596a43ea3d06eec93b7a189ff0