Implement q_sort
Sort the elements of the queue using merge sort in ascending or descending order.
First, separate the head from the first element of the queue. Second, use the
mergesort_list function to split the list into two at the middle, recursively
until the head is NULL or the list is empty. Third, use mergeTwoLists to
merge each two list into a complete list. Finally, connect the head to the first
element of the queue. If the head is NULL, the queue is empty, or there is
only one node, do nothing.

Change-Id: I45e3ae507780c9c864a955325182cf8bceb58f53