Implement q_ascend
Sort the elements of the queue in ascending order and delete every node
that has a node with a strictly greater value anywhere to its right. First,
use the q_sort function to sort the queue in ascending order. Second,
iterate through the nodes in the queue. If a node's value, compared to the
next node's value using strcmp, is greater than or equal to 0, delete
the node. Finally, return the number of nodes deleted. If the head is NULL,
the queue is empty, or there is only one node, return 0.

Change-Id: I588f8aaebe6363ab4247c6e97d80ff63bda43561