Implement q_descend
Sort the elements of the queue in descending order and delete every node
that has a node with a strictly lesser value anywhere to its right. First,
use the q_sort function to sort the queue in descending order. Second,
iterate through the nodes in the queue. If a node's value, compared to the
next node's value using strcmp, is less than or equal to 0, delete the
node. Finally, return the number of nodes in the queue. If the head is
NULL or the queue is empty, return 0. If there is only one node, return 1.

Change-Id: I8d72f00b02e2687bfc4303639e078e1b6cbb6aeb