Implement q_merge
Merge all queues into one sorted queue, either in ascending or descending order.
First, iterate through the list of chain to find all queues. Second, separate
the heads of the two queues that will be merged. Third, use the mergeTwoLists
function to merge the two queues into the first queue. Finally, connect the
head to the new queue. After iterating through all queues in the chain list,
iterate through the first queue to count the number of nodes and return the count.
If the head is NULL or the queue is empty, return 0. If there is only one
node, return 1.

Change-Id: I1d3c8b003ba3b6c668afa4d8f4aa5097628fce75